% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_functions.R
\name{avg_neighbor_distances}
\alias{avg_neighbor_distances}
\title{Estimate quality of mapped data based on distances of between reference cells}
\usage{
avg_neighbor_distances(
  query_seurat_object,
  reference_seurat,
  reduction_name = "scvi",
  query_nn = "query_ref_nn",
  distance.metric = "cosine",
  add_to_seurat = TRUE
)
}
\arguments{
\item{query_seurat_object}{query seurat}

\item{reference_seurat}{reference seurat}

\item{reduction_name}{name of (scvi) reduction in reference.}

\item{query_nn}{character vector with name of nn object in query (that conatins neighbor indices in reference)}

\item{distance.metric}{distance metric, 'cosine' or any available for stats::dist like 'euclidean'}

\item{add_to_seurat}{add result to query seurat object ? defualts to TRUE}
}
\value{
query_seurat_object with 'avg_neighbor_distance' in meta.data or vector
}
\description{
Calculates average distance between reference neighbors.
}
