% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_functions.R
\name{check_distance_neighbors}
\alias{check_distance_neighbors}
\title{Estimate quality of mapped data based on distances of query cells from reference cells}
\usage{
check_distance_neighbors(
  query_seurat_object,
  reference_seurat,
  reduction_name = "scvi",
  query_nn = "query_ref_nn",
  reference_nn = "reference_nn",
  k_param = 30,
  annoy.metric = "cosine",
  global_seed = 12345
)
}
\arguments{
\item{query_seurat_object}{Seurat object}

\item{reference_seurat}{reference seurat}

\item{reduction_name}{name of scvi reduction in reference AND query. Also expect a umap named as 'umap_'reference_reduction}

\item{query_nn}{character vector with name}

\item{reference_nn}{character vector with name}

\item{k_param}{k neighbors}

\item{annoy.metric}{string passed to Seurat's FindNeighbors}

\item{global_seed}{seed}
}
\value{
query_seurat_object with quality results in metadata
nn_distance_reference: what is the median distance of a query cell to all neighbor cells in reference
normalized_nn_distance: nn_distance_reference / median_dist_refNeighbors (how comparable are the query cells distances to their neighboring cell distances)
}
\description{
Calculates average distance to reference neighbors and normalized average distances based on average distances of these reference neighbors within the reference dataset.
}
