% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_functions.R
\name{check_freq_neighbors}
\alias{check_freq_neighbors}
\title{Estimate quality of mapped data}
\usage{
check_freq_neighbors(
  query_seurat_object,
  reference_seurat,
  reduction_name = "scvi",
  k_param = 30,
  annoy.metric = "cosine",
  global_seed = 12345
)
}
\arguments{
\item{query_seurat_object}{Seurat object}

\item{reference_seurat}{reference seurat}

\item{reduction_name}{name of scvi reduction in reference AND query. Also expect a umap named as 'umap_'reference_reduction}

\item{k_param}{k neighbors}

\item{annoy.metric}{'cosine' or 'euclidean'}

\item{global_seed}{seed}
}
\value{
query_seurat_object with quality results in metadata
query_neighbor_pct: how many neighbors are from query dataset. Can either indicate that query contains many more cells than reference or that celltype is absent in reference.
}
\description{
Calculates a shared KNN to see how well query cells mix with reference cells and how the distances are.
}
