% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{check_quality_markers}
\alias{check_quality_markers}
\title{Estimate quality of mapped data based on marker genes}
\usage{
check_quality_markers(
  query_seurat_object,
  reference_seurat,
  reduction_name = "scvi",
  k_param = 30,
  global_seed = 12345
)
}
\arguments{
\item{query_seurat_object}{Seurat object}

\item{reference_seurat}{reference seurat}

\item{reduction_name}{name of scvi reduction in reference AND query. Also expect a umap named as 'umap_'reference_reduction}

\item{global_seed}{seed}
}
\value{
query_seurat_object with quality results in metadata
}
\description{
Check marker gene expression in query cells based on predicted clusters
}
