% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{compare_clustering}
\alias{compare_clustering}
\title{Compare two clustering results}
\usage{
compare_clustering(
  query_seura_object,
  clustering_1,
  clustering_2,
  min_cells = 10,
  min_pct = 0.1,
  return_data = FALSE
)
}
\arguments{
\item{query_seura_object}{query seurat object}

\item{clustering_1}{column name in query_seura_object metadata with cluster labels}

\item{clustering_2}{column name in query_seura_object metadata with cluster labels}

\item{min_cells}{minimum shared cells to construct a relationship (edge) between two clusters from different clusterings}

\item{min_pct}{min percentage to construct a relationship (edge) between two clusters from different clusterings}

\item{return_data}{return data or  plot. defaults to FALSE which means the plot will be printed}
}
\value{
ggplot or data behind plot
}
\description{
For example between propagated and original clusters!
TODO: add description
}
