% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{convert_to_seurat}
\alias{convert_to_seurat}
\title{Convert various objects to seurat object.}
\usage{
convert_to_seurat(
  object,
  suffix = "query",
  metadata = NULL,
  global_seed = 12345
)
}
\arguments{
\item{object}{object that will be converted to seurat. For matrix: expects columns as cells.}

\item{suffix}{query project name. defaults to 'query'}

\item{global_seed}{seed}

\item{metadata:}{optional metadata data.frame. rownames must correspond to colnames of object}
}
\value{
formatted seurat object
}
\description{
Helper function to prepare seurat object from matrix, SingleCellObject or ...
}
