% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lamanno_seurat}
\alias{lamanno_seurat}
\title{Dopaminergic neuron single cell data.}
\format{
A Seurat object
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE76381}
}
\usage{
lamanno_seurat
}
\description{
From the the La Manno brain data ("mouse-adult") loaded via scRNAseq::LaMannoBrainData.
}
\keyword{datasets}
