% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_cluster_tree}
\alias{plot_cluster_tree}
\title{Plot metadata entries in cluster tree using ggtree}
\usage{
plot_cluster_tree(treedata, metadata, leaf_level)
}
\arguments{
\item{treedata}{TODO}

\item{metadata}{TODO}

\item{leaf_level}{which level to use as leaves ?}
}
\value{
ggtree object or plot
}
\description{
TODO: add description
TODO: add reference to ggtree vignette
}
\examples{


}
