% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_propagation_stats}
\alias{plot_propagation_stats}
\title{Plot propagation statistics about query column}
\usage{
plot_propagation_stats(
  query_seura_object,
  reference_seurat,
  label_col,
  label_col_query = "predicted",
  return_data = FALSE
)
}
\arguments{
\item{query_seura_object}{query seurat object}

\item{reference_seurat}{reference seurat object}

\item{label_col}{the column name in reference_map metadata with labels to propagate}

\item{label_col_query}{he column name in query_seura_object metadata with labels to propagate}

\item{return_data}{return data and / or  plot}
}
\value{
plot or data behind plot
}
\description{
Determines over or und represented clusters
TODO: add description
}
\examples{


}
