% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_query_labels}
\alias{plot_query_labels}
\title{Plot labels on query and reference to compare visually.}
\usage{
plot_query_labels(
  query_seura_object,
  reference_seurat,
  label_col,
  label_col_query = "predicted",
  overlay = FALSE,
  bg_col = "grey80",
  overlay_color = "red",
  overlay_alpha = 0.5,
  query_pt_size = NULL,
  query_umap = "umap_scvi",
  reference_umap = "umap_scvi",
  labelonplot = TRUE,
  cols_plot = NULL,
  noaxes = TRUE,
  nolegend = TRUE,
  ...
)
}
\arguments{
\item{query_seura_object}{query seurat object}

\item{reference_seurat}{reference seurat object}

\item{label_col}{the column name in reference_map metadata with labels to propagate}

\item{label_col_query}{he column name in query_seura_object metadata with labels to propagate}

\item{overlay}{overlay query onto label}

\item{bg_col}{if non null, the reference will take this color and the query ill be overlayed colored by label_col_query. This will ignore overlay_color and overlay_alpha}

\item{overlay_color}{color of overlay points}

\item{overlay_alpha}{= alpha for overlay plot}

\item{query_pt_size}{numeric for pt size of query cells. defaults to NULL which will inherit the pt.size from the reference DimPlot}

\item{query_umap}{name of query umap. defaults to "umap_scvi"}

\item{reference_umap}{name of reference umap. defaults to "umap_scvi"}

\item{labelonplot}{put labels on plot. defaults to TRUE}

\item{cols_plot}{passed to Seurat::DimPlot cols argument. Defaults to NULL}

\item{noaxes}{don't plot axes on UMAP. defaults to TRUE}

\item{nolegend}{don't plot legend. defaults to TRUE}

\item{...}{additional arguments to Seurat::DimPlot}
}
\value{
plot
}
\description{
Is a wrapper around Seurat::DimPlot functions that allows to plot qeury and reference side-by-side or on top of each other
The overlay parameters allow to change the appearance of query points and the alpha of the reference points.
}
\examples{


}
