% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{predict_query}
\alias{predict_query}
\title{Predict latent space of embedding}
\usage{
predict_query(
  query_seurat_object,
  model_path,
  query_reduction = "scvi",
  max_epochs = 30,
  assay = "RNA",
  global_seed = 12345
)
}
\arguments{
\item{query_seurat_object}{Seurat object}

\item{model_path}{path to pretrained scvi model}

\item{query_reduction}{name for reduction}

\item{max_epochs}{epochs to train}

\item{assay}{assay name for new prediction and where to find raw counts. defaults to RNA}

\item{global_seed}{seed}
}
\value{
query_seurat_object: Updated seurat object with predicted latent space as reduction.
}
\description{
Function to run scvi model prediction on query seurat based on provided scvi model.
Uses the scArches algorithm as described in the documentation: https://docs.scvi-tools.org/en/stable/api/reference/scvi.model.SCVI.load_query_data.html
}
