% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{prepare_query_hypoMap}
\alias{prepare_query_hypoMap}
\title{Prepare seurat object for mapping onto hypoMap}
\usage{
prepare_query_hypoMap(
  object,
  suffix = "query",
  covariates = c(batch_var = "Batch_ID", inferred_sex = "inferred_sex.x",
    rpl_signature_expr_median = "rpl_signature_expr_median"),
  sex_var = "Sex",
  ...
)
}
\arguments{
\item{object}{input object: a Seurat, SingleCellExperiment or matrix-like object containing the raw (non-normalized) counts.}

\item{suffix}{query project name. defaults to 'query'}

\item{covariates}{named vector: names of covariates for hypoMap scvi model that will be added via this function. Changing this parameter or supplying the variables without the proper format can cause problems! defaults to: c(batch_var = "Batch_ID",inferred_sex = "inferred_sex.x",rpl_signature_expr_median = "rpl_signature_expr_median")}

\item{sex_var}{column name with annotation of sample sex}

\item{...}{additional arguments to prepare_query}
}
\value{
updated seurat object
}
\description{
Function to prepare an object for embedding with relevant covariates and metadata columns to work with HypoMap.
This will call prepare_query and additionally check the covariates and add them with a custom procedure (most likely does not work for other scvi models and reference objects)
}
