% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{project_query}
\alias{project_query}
\title{Project query into reference UMAP}
\usage{
project_query(
  query_seurat_object,
  reference_map_reduc,
  reference_map_umap,
  query_reduction = "scvi",
  assay = "RNA",
  use_projectUMAP = FALSE,
  annoy.metric = "cosine",
  result_type = "label",
  label_vec = NULL,
  global_seed = 12345
)
}
\arguments{
\item{query_seurat_object}{Seurat object}

\item{reference_map_reduc}{reference map dim_reduc (scvi latent space)}

\item{reference_map_umap}{reference map UMAP dim_reduc}

\item{query_reduction}{key of predicted reduction used for UMAP projection}

\item{assay}{assay name. defaults to RNA}

\item{use_projectUMAP}{whether to use Seurat's projectUMAP  or a manual adaption (circumventing possible problems between Seurat and uwot)}

\item{annoy.metric}{'cosine' or 'euclidean'}

\item{result_type}{return type: "all","label","entropy" .  For label the maximum is taken to return the likeliest assignment. All returns all results (multiple colummns!), entropy calculates the entropy across probabilities as an uncertainty measure but does not return a label}

\item{label_vec}{a vector with labels from reference that will be propagated to query (requires same order as query_reduction!). defaults to NULL (nothing will be propagated). See also 'propagate_labels'}

\item{global_seed}{seed}

\item{n_neighbors}{n neighbors passed to ProjectUMAP}
}
\value{
query_seurat_object: Updated seurat object with projected UMAP, labels and knn graph
}
\description{
Function to project query onto reference using a predicted scvi latent space. Returns UMAP, propagated labels and kNN in the updated seurat object.
}
