% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{propagate_labels}
\alias{propagate_labels}
\title{Propagate labels to query}
\usage{
propagate_labels(nn_idx, label_vec)
}
\arguments{
\item{nn_idx}{kNN}

\item{label_vec}{vector with labels, ids must be consistent with NN indices in kNN}
}
\value{
predicted label for each cell (kNN row)
}
\description{
Use kNN to propagate labels (any metadata column from reference) to query.
}
