% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_functions.R
\name{propagate_labels_prob}
\alias{propagate_labels_prob}
\title{Estimate quality of mapped data based on marker genes from reference}
\usage{
propagate_labels_prob(
  query_seurat_object,
  reference_seurat_object,
  label_col,
  reduction_name_query = "scvi",
  reduction_name_reference = "scvi",
  k.param = 30,
  add_entropy = FALSE,
  add_to_seurat = TRUE
)
}
\arguments{
\item{reference_seurat_object}{v}

\item{add_to_seurat}{}
}
\value{
object
}
\description{
Cell probabilities similar to scARches algorithm
This function runs per cell
}
