% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{query_romanov}
\alias{query_romanov}
\title{Hypothalamus neuron single cell data from Romanov et al.}
\format{
A Seurat object
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE74672}
}
\usage{
query_romanov
}
\description{
Manually curated data as ax example datasets for mapping new data.
Comapred to the full smart-seq datasets this dataset is already reduced to potential neurons to demonstrate mapping on the neuron map in the readme!
}
\keyword{datasets}
