% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{reconstruct_levels}
\alias{reconstruct_levels}
\title{Reconstruct high cluster levels based on provided edgelist}
\usage{
reconstruct_levels(
  edgelist,
  input_annotation,
  level_prefix = "K[0-9]+",
  result_prefix = ""
)
}
\arguments{
\item{edgelist}{dataframe or matrix with two columns (edgelist: to and from). optional third column giving the 'level' of 'from' which will be used to determine reconstructed levels and name columns}

\item{input_annotation}{query project name. defaults to 'query'}

\item{level_prefix}{regex used to extract 'level' if third column is missing}

\item{result_prefix}{prefix to add to result column names}
}
\value{
dataframe with columns for input_annotation and all clustering levels (above) in edgelist
}
\description{
After prediction or label propagation using a low clustering level this function can be used to reconstruct higher clustering levels.
Is a wrapper around 'recursive_reconstruction'
Requires a clustertree defined in a simple edgelist. User has to ensure that provided input is a valid edgelist!
}
