% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{recursive_reconstruction}
\alias{recursive_reconstruction}
\title{Reconstruct high cluster levels based on provided edgelist}
\usage{
recursive_reconstruction(edgelist, input_annotation)
}
\arguments{
\item{edgelist}{dataframe or matrix with two columns (edgelist: to and from). optional third column giving the 'level' of 'from' which will be used to determine reconstructed levels and name columns}

\item{input_annotation}{query project name. defaults to 'query'}
}
\value{
matrix with columns for input_annotation and all clustering levels (above) in edgelist
}
\description{
Recursively go through edgelist and reconstruct higher-level annotation of input_annotation
}
