% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reference_hypoMap_full}
\alias{reference_hypoMap_full}
\title{Reference object for mapping of new hypothalamus data}
\format{
Seurat object with metadata and reductions 'scvi' and 'umap_scvi' including the umap model.
}
\source{
\url{placeholder}
}
\usage{
reference_hypoMap_full
}
\description{
This is a simplified object of the full HypoMap with all celltypes, without count values to reduce the size.
}
\keyword{datasets}
