% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_lamanno_da}
\alias{sce_lamanno_da}
\title{Dopaminergic neuron single cell data.}
\format{
A SingleCellExperiment object:
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE76381}
}
\usage{
sce_lamanno_da
}
\description{
From the the La Manno brain data ("mouse-adult") loaded via scRNAseq::LaMannoBrainData.
}
\keyword{datasets}
