% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{transform_seurat}
\alias{transform_seurat}
\title{Update gene ids}
\usage{
transform_seurat(query_seurat_object, suffix = "query", global_seed = 12345)
}
\arguments{
\item{query_seurat_object}{Seurat object}

\item{suffix}{project name to clearly label various steps. defaults to 'query'}

\item{global_seed}{seed}
}
\value{
formatted seurat object
}
\description{
Helper function to prepare seurat object gene ids for format used in reference
Currently does nothing and just returns the same object.
}
