% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{visNetwork_clustering}
\alias{visNetwork_clustering}
\title{Compare two clustering results with visNetwork}
\usage{
visNetwork_clustering(
  query_seura_object,
  clustering_1,
  clustering_2,
  min_cells = 10,
  min_pct = 0.1,
  text_size = 25,
  px_height = "800px",
  col1 = "#cc2118",
  col2 = "#302ac9",
  color_by = NULL,
  palette = "Set3",
  return_data = FALSE
)
}
\arguments{
\item{query_seura_object}{query seurat object}

\item{clustering_1}{column name in query_seura_object metadata with cluster labels}

\item{clustering_2}{column name in query_seura_object metadata with cluster labels}

\item{min_cells}{minimum shared cells to construct a relationship (edge) between two clusters from different clusterings}

\item{min_pct}{min percentage to construct a relationship (edge) between two clusters from different clusterings}

\item{text_size}{numeric, font size in sankey}

\item{px_height}{height of interactve visNetwork in pixles}

\item{col1}{string, color name for clustering_1}

\item{col2}{string, color name for clustering_2}

\item{color_by}{defaults to NULL. if not NULL a input_clusters column name to color edges by}

\item{palette}{= Rcolorbrewer palette name}

\item{return_data}{return data or  plot. defaults to FALSE which means the plot will be printed}
}
\value{
ggplot or data behind plot
}
\description{
For example between propagated and original clusters!
TODO: add description
}
