% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{adjusted_cell_probabilities}
\alias{adjusted_cell_probabilities}
\title{Calculate probabilities for labels of neighboring cells}
\usage{
adjusted_cell_probabilities(
  dist_Nc,
  labels_of_neighbor,
  with_euclidean = FALSE,
  result_type = "all"
)
}
\arguments{
\item{dist_Nc}{vector of length k with euclidean distances to neigbors}

\item{labels_of_neighbor}{vector of length k with labels of neighbors}

\item{with_euclidean}{if input is euclidean: use gussian smoothing and convert to similarity through this. If not assume cosine distance and just invert using: 1 - dist_Nc^2/2}

\item{result_type}{return type: "all","label","entropy" .  For label the maximum is taken to return the likeliest assignment. All returns all results (multiple colummns!), entropy calculates the entropy across probabilities as an uncertainty measure but does not return a label}
}
\value{
label probability depending on result_type
}
\description{
Cell probabilities similar to scARches algorithm
This function runs per cell.
}
