% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{adjusted_cell_probabilities}
\alias{adjusted_cell_probabilities}
\title{Estimate quality of mapped data based on marker genes from reference}
\usage{
adjusted_cell_probabilities(
  dist_Nc,
  labels_of_neighbor,
  apply_gaussian = TRUE,
  result_type = "all"
)
}
\arguments{
\item{dist_Nc}{vector of length k with euclidean distances to neigbors}

\item{result_type}{return type: "all","label","entropy"}

\item{label_of_neighbor}{vector of length k with labels of neighbors}
}
\value{
label probability depending on result_type
}
\description{
Cell probabilities similar to scARches algorithm
This function runs per cell
}
