% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_functions.R
\name{avg_neighbor_distances}
\alias{avg_neighbor_distances}
\title{Estimate quality of mapped data based on distances of between reference cells}
\usage{
avg_neighbor_distances(
  neighbors_object = NULL,
  reference_map_reduc,
  query_seurat_object = NULL,
  query_nn = "query_ref_nn",
  distance.metric = "cosine",
  add_to_seurat = TRUE
)
}
\arguments{
\item{neighbors_object}{neighbors_object. if NULL tries to find it in query_seurat_object}

\item{reference_map_reduc}{latent space from reference with same order as idx in \code{neighbors_object@nn.idx}}

\item{query_seurat_object}{query seurat. defaults to NULL}

\item{query_nn}{character vector with name of nn object in query (that contains neighbor indices in reference). only used when neighbors_object is NULL}

\item{distance.metric}{distance metric, 'cosine' or any available for stats::dist like 'euclidean'}

\item{add_to_seurat}{add result to query seurat object ? requires query_seurat_object to be  defaults to FALSE}
}
\value{
query_seurat_object with 'avg_neighbor_distance' in meta.data or vector
}
\description{
Calculates average distance between reference neighbors.
}
