% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{map_new_seurat_hypoMap}
\alias{map_new_seurat_hypoMap}
\title{Map a query seurat onto a reference}
\usage{
map_new_seurat_hypoMap(
  query_seurat_object,
  suffix = "query",
  subset_col = "",
  label_col = "",
  subset_values = NULL,
  max_epochs,
  reference_map_reduc = NULL,
  reference_map_umap = NULL,
  reference_map_metadata = NULL,
 
    model_path = "/beegfs/scratch/bruening_scratch/lsteuernagel/data/scHarmonize/hypothalamusMapNeurons_v4/harmonization_results/hypothalamus_neurons_reference/hypothalamus_neurons_reference_model/",
  use_reticulate = FALSE,
  temp_dir = "/beegfs/scratch/bruening_scratch/lsteuernagel/data/tmp_mapscvi/",
 
    dim_red_file = "/beegfs/scratch/bruening_scratch/lsteuernagel/data/tmp_mapscvi/hypoMap_dimred.RData",
  global_seed = 12345
)
}
\arguments{
\item{query_seurat_object}{Seurat object}

\item{suffix}{project name to clearly label various steps. defaults to 'query'}

\item{subset_col}{character value: metadata column in query_seurat_object that will be used to subset data. defaults to ''}

\item{label_col}{the column name in reference_map_metadata with labels to propagate}

\item{subset_values}{character vector: which values from subset_col should be kept. defaults to NULL which will prevent any subsetting}

\item{max_epochs}{epochs to train}

\item{reference_map_reduc}{reference map dim_reduc (scvi latent space)}

\item{reference_map_umap}{reference map UMAP dim_reduc}

\item{reference_map_metadata}{ref meta dataframe}

\item{model_path}{path to pretrained scvi model}

\item{use_reticulate}{if TRUE: tries to call scvi via reticulate, if FALSE: exports an anndata object to a temp file and then calls a python script to run prediction code using system()}

\item{temp_dir}{directory for temporary files if use_reticulate == FALSE}

\item{dim_red_file}{RData with hypoMap dimred}

\item{global_seed}{seed}
}
\value{
formatted seurat object
}
\description{
Wrapper functions that executes low level functions to prepare, predict and project new data
}
