% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{map_new_seurat_hypoMap}
\alias{map_new_seurat_hypoMap}
\title{Map a query seurat onto a reference}
\usage{
map_new_seurat_hypoMap(
  query_seurat_object,
  suffix = "query",
  subset_col = "",
  label_col = "",
  subset_values = NULL,
  max_epochs,
  reference_map_reduc = NULL,
  reference_map_umap = NULL,
  reference_map_metadata = NULL,
 
    model_path = "/beegfs/scratch/bruening_scratch/lsteuernagel/data/scHarmonize/hypothalamusMapNeurons_v4/harmonization_results/hypothalamus_neurons_reference/hypothalamus_neurons_reference_model/",
  global_seed = 12345
)
}
\arguments{
\item{query_seurat_object}{Seurat object}

\item{suffix}{project name to clearly label various steps. defaults to 'query'}

\item{subset_col}{column to subset by}

\item{subset_values}{values of subset_col to filter on}

\item{max_epochs}{epochs to train during scvi query function}

\item{model_path}{path to scvi model on disk}

\item{global_seed}{seed}
}
\value{
formatted seurat object
}
\description{
Wrapper functions that executes low level functions to prepare, predict and project new data
}
