% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{map_new_seurat_hypoMap}
\alias{map_new_seurat_hypoMap}
\title{Map a query seurat onto the hypomap reference}
\usage{
map_new_seurat_hypoMap(
  query_seurat_object,
  suffix = "query",
  assay = "RNA",
  subset_col = "",
  label_col = "",
  subset_values = NULL,
  max_epochs,
  model_path = system.file("extdata/models/hypoMap_harmonized_scVI_model/", package =
    "mapscvi"),
  reference_seurat = mapscvi::reference_hypoMap_downsample,
  reference_reduction = "scvi",
  inferred_sex_varname = "inferred_sex",
  pythonEnv = NULL,
  use_reticulate = FALSE,
  global_seed = 12345
)
}
\arguments{
\item{query_seurat_object}{Seurat object}

\item{suffix}{query project name. defaults to 'query'}

\item{assay}{which assay from query_seurat_object. defaults to RNA}

\item{subset_col}{character value: metadata column in query_seurat_object that will be used to subset data. defaults to ''}

\item{label_col}{the column name in reference_map_metadata with labels to propagate}

\item{subset_values}{character vector: which values from subset_col should be kept. defaults to NULL which will prevent any subsetting}

\item{max_epochs}{epochs to train}

\item{model_path}{path to pretrained scvi model}

\item{reference_seurat}{reference seurat}

\item{reference_reduction}{name of scvi reduction in reference. Also expect a umap named as 'umap_'reference_reduction}

\item{inferred_sex_varname}{variable name for inferred sex that has to correspond to model (will automatically be set when using reference_mode)}

\item{pythonEnv}{path to a python binary, passed to reticulate::use_python}

\item{use_reticulate}{if TRUE: tries to call scvi via reticulate, if FALSE: exports an anndata object to a temp file and then calls a python script to run prediction code using system()}

\item{global_seed}{seed}

\item{reference_mode}{directly specify what to use as reference. Possible values are either 'hypoMap_neurons', 'hypoMap_full' or 'manual'. This will set default values for reference_seurat, reference_reduction and model_path that can be overwritten by directly specifying them. Setting this parameter to 'manual' or NULL requires valid entries for the other parameters.}
}
\value{
formatted seurat object
}
\description{
Wrapper functions that executes low level functions to prepare, predict and project new data onto a reference based on an scvi model.
Defaults to a smaller object (that can be shipped with the package) for reference_seurat (mapscvi::reference_hypoMap_downsample). Please alternatively download and specify the full object: https://www.repository.cam.ac.uk/handle/1810/341048
If the model differs strongly from the hypoMap model (covariates used etc) the prepare and mapping functions might break.
In that case please manually run the prepare, predict and project functions as shown in the readme.
}
