% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_propagation_stats}
\alias{plot_propagation_stats}
\title{Plot propagation statistics about query column}
\usage{
plot_propagation_stats(
  query_seura_object,
  reference_seurat,
  label_col_query,
  label_col,
  ratio_cut = 2,
  text_size = 10,
  plot_abs = FALSE,
  n_top = 10,
  ref_col = "#eda09a",
  query_col = "#9aa9ed",
  return_data = FALSE
)
}
\arguments{
\item{query_seura_object}{query seurat object}

\item{reference_seurat}{reference seurat object}

\item{label_col_query}{he column name in query_seura_object metadata with labels to propagate}

\item{label_col}{the column name in reference_map metadata with labels to propagate}

\item{ratio_cut}{what ratio is considered to be a relevant difference (just for coloring)}

\item{text_size}{text_size in ggplots}

\item{plot_abs}{use absolute cell counts in barplot (might be option for large query data). defaults to FALSE}

\item{n_top}{how many clusters to include in barplots}

\item{ref_col}{color in plot}

\item{query_col}{color in plot}

\item{return_data}{return data or  plot. defaults to FALSE which means the plot will be printed}
}
\value{
ggplot or data behind plot
}
\description{
Determines over or und represented clusters
TODO: add description
}
\examples{


}
