% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_sankey_comparison}
\alias{plot_sankey_comparison}
\title{Plot comparison between reference and query using sankey plots}
\usage{
plot_sankey_comparison(metadata, return_data = FALSE)
}
\arguments{
\item{metadata}{TODO}

\item{return_data}{}
}
\value{
ggtree object or plot
}
\description{
TODO: add description
}
\examples{


}
