% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{prepare_query}
\alias{prepare_query}
\title{Prepare seurat object for mapping functions}
\usage{
prepare_query(
  query_seurat_object,
  suffix = "query",
  assay = "RNA",
  subset_col = "",
  subset_values = NULL,
  normalize = TRUE,
  global_seed = 12345
)
}
\arguments{
\item{query_seurat_object}{Seurat object}

\item{suffix}{project name to clearly label various steps. defaults to 'query'}

\item{assay}{which assay from query_seurat_object. defaults to RNA}

\item{subset_col}{character value: metadata column in query_seurat_object that will be used to subset data. defaults to ''}

\item{subset_values}{character vector: which values from subset_col should be kept. defaults to NULL which will prevent any subsetting}

\item{normalize}{boolean: run standard normalization on assay. defaults to TRUE}

\item{global_seed}{seed}
}
\value{
updated seurat object
}
\description{
Generic Function to prepare a Seurat object for embedding
}
