% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{project_query}
\alias{project_query}
\title{Project query into reference UMAP}
\usage{
project_query(
  query_seurat_object,
  reference_map_reduc,
  reference_map_umap,
  query_reduction = "scvi",
  assay = "RNA",
  k_param_umap = 30,
  annoy.metric = "cosine",
  label_vec = NULL,
  global_seed = 12345
)
}
\arguments{
\item{query_seurat_object}{Seurat object}

\item{reference_map_reduc}{reference map dim_reduc (scvi latent space)}

\item{reference_map_umap}{reference map UMAP dim_reduc}

\item{query_reduction}{key of predicted reduction used for UMAP projection}

\item{assay}{assay name. defaults to RNA}

\item{k_param_umap}{k neighbors passed to ProjectUMAP}

\item{annoy.metric}{'cosine' or 'euclidean'}

\item{label_vec}{a vector with labels from reference that will be propagated to query (requires same order as query_reduction!). defaults to NULL (nothing will be propagated). See also 'propagate_labels'}

\item{global_seed}{seed}
}
\value{
query_seurat_object: Updated seurat object with projected UMAP, labels and knn graph
}
\description{
Function to project query onto reference using a predicted scvi latent space. Returns UMAP, propagated labels and kNN in the updated seurat object.
}
