% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{update_gene_names}
\alias{update_gene_names}
\title{Update gene ids}
\usage{
update_gene_names(
  query_seurat_object,
  reference_seurat_object = NULL,
  reference_gene_identifiers = NULL,
  query_species = NULL,
  reference_species = NULL
)
}
\arguments{
\item{query_seurat_object}{query Seurat object}

\item{reference_seurat_object}{reference_seurat_object. Defaults to NULL. Then reference_gene_identifiers will be used}

\item{reference_gene_identifiers}{a character vector of gene identifiers}

\item{query_species}{species}

\item{reference_species}{species}
}
\value{
formatted seurat object
}
\description{
Helper function to prepare seurat object gene ids for format used in reference
Currently does nothing and just returns the same object.
}
